requirements.txt
Flask
pycryptodome
werkzeug